preserve exit code when running pipes.
Ignore-this: 27351c08b537b9dbc18486caf427be49

This fixes a bug where

echo foo | grep bar && echo found it

was giving the wrong result, because we were ignoring the exit code
that was output by pipes.  Unfortunately, it's rather an invasive
change.  I can't help but wonder whether we'd be better off storing
the exit code in the state? In "real" shells, the exit status of the
last command is stored in the environment variable $?, so it seems
reasonable to mimic this behavior.

darcs-hash:20081119211458-72aca-e36045a6e864431db49eff972b3f43ee911daa37