Better algorithm for pipes and redirects
Ignore-this: 245bbf0238b2b8bf360033d1c01d2958
Basically, and computation that can possibly be part of a pipeline
gets a new type, which is a function (InputPipe -> Shell ExitCode).
The InputPipe may contain a pair of MVar's, one for the computation to
send back a WriteHandle, and the other to send back the ExitCode when
it's done.  The goal is to not even really export InputPipe, except
that then we need some way to chain Shell's together, passing it
along.  But certainly we needn't export the constructor.  This still
needs some cleaning up, but it seems to work efficiently in all
combinations of builtins/external processes (i.e. only creating a Chan
when necessary).

darcs-hash:20081122201107-0087a-a3943a7a411632d217d38143566cde1831eea5dd