fix two bugs in mkdir
Ignore-this: 5aac8a1d54181a2ae4311554259edd45
One was that when mkdir -p was given a relative path, it could loop forever
making "", so long as createDirectory "" succeeds and doesDirectoryExist ""
fails (which seems to be the case in ghc 6.6).  The other was that mkdir -p
isn't supposed to complain if the directory already exists.

darcs-hash:20081125131924-72aca-3486729d0a602983ca811eba66668be38a52b120